pyproject.toml
[tool.poetry]
name = "nonebot-plugin-error-alert"
version = "0.2.0"
description = ""
authors = ["ssttkkl <huang.wen.long@hotmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "nonebot_plugin_error_alert"}]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
nonebot2 = "^2.0.0"
nonebot-plugin-apscheduler = ">=0.3.0"
nonebot-plugin-send-anything-anywhere = ">=0.3.0"
nonebot-plugin-datastore = ">=1.0.0"
nb-cli = "^1.2.3"


[tool.poetry.group.dev.dependencies]
nonebot2 = {extras = ["websockets"], version = "^2.0.1"}
nonebot-adapter-qqguild = "^0.2.5"
nonebot-plugin-escape-url = "^0.1.3.post1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
